third_party/WebKit/LayoutTests/fast/html/imports/import-nested-dup-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M73, around March 2019. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 for more details.
Test that a subimport should be loaded even if is linked as a top-level import.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks[i] is expected[i]
PASS marks.length is expected.length
PASS successfullyParsed is true

TEST COMPLETE
